Adapting local deepnet predictions for text and items fields
