+++ bigml/modelfields.py
@@ -24,10 +24,12 @@ is used for local predictions.
 
 """
 import logging
+import re
 
 from bigml.util import invert_dictionary, DEFAULT_LOCALE
 from bigml.fields import DEFAULT_MISSING_TOKENS
 from bigml.resourcehandler import get_resource_type
+from bigml.predicate import TM_FULL_TERM, TM_ALL
 
 
 LOGGER = logging.getLogger('BigML')
@@ -39,6 +41,28 @@ FIELDS_PARENT = { \
     "deepnet": "deepnet"}
 
 
+def parse_terms(text, case_sensitive=True):
+    """Returns the list of parsed terms
+
+    """
+    if text is None:
+        return []
+    expression = ur'(\b|_)([^\b_\s]+?)(\b|_)'
+    pattern = re.compile(expression)
+    return [match[1] if case_sensitive else match[1].lower()
+            for match in re.findall(pattern, text)]
+
+
+def parse_items(text, regexp):
+    """Returns the list of parsed items
+
+    """
+    if text is None:
+        return []
+    pattern = re.compile(regexp, flags=re.U)
+    return pattern.split(text)
+
+
 def check_model_structure(model, inner_key="model"):
     """Checks the model structure to see if it contains all the
     main expected keys
@@ -48,6 +72,31 @@ def check_model_structure(model, inner_key="model"):
             model['resource'] is not None and
             ('object' in model and inner_key in model['object'] or
              inner_key in model))
+
+
+def get_unique_terms(terms, term_forms, tag_cloud):
+    """Extracts the unique terms that occur in one of the alternative forms in
+       term_forms or in the tag cloud.
+
+    """
+
+    extend_forms = {}
+    for term, forms in term_forms.items():
+        for form in forms:
+            extend_forms[form] = term
+        extend_forms[term] = term
+    terms_set = {}
+    for term in terms:
+        if term in tag_cloud:
+            if term not in terms_set:
+                terms_set[term] = 0
+            terms_set[term] += 1
+        elif term in extend_forms:
+            term = extend_forms[term]
+            if term not in terms_set:
+                terms_set[term] = 0
+            terms_set[term] += 1
+    return terms_set.items()
 
 
 def check_model_fields(model):
@@ -87,7 +136,8 @@ class ModelFields(object):
     """
 
     def __init__(self, fields, objective_id=None, data_locale=None,
-                 missing_tokens=None):
+                 missing_tokens=None, terms=False, categories=False,
+                 numerics=False):
         if isinstance(fields, dict):
             try:
                 self.objective_id = objective_id
@@ -101,8 +151,51 @@ class ModelFields(object):
                     self.data_locale = DEFAULT_LOCALE
                 if self.missing_tokens is None:
                     self.missing_tokens = DEFAULT_MISSING_TOKENS
+                if terms:
+                    # adding text and items information to handle terms
+                    # expansion
+                    self.term_forms = {}
+                    self.tag_clouds = {}
+                    self.term_analysis = {}
+                    self.items = {}
+                    self.item_analysis = {}
+                if categories:
+                    self.categories = {}
+                if terms or categories or numerics:
+                    self.add_terms(categories, numerics)
+
             except KeyError:
                 raise Exception("Wrong field structure.")
+
+    def add_terms(self, categories=False, numerics=False):
+        """Adds the terms information of text and items fields
+
+        """
+        for field_id, field in self.fields.items():
+            if field['optype'] == 'text':
+                self.term_forms[field_id] = {}
+                self.term_forms[field_id].update(
+                    field['summary']['term_forms'])
+                self.tag_clouds[field_id] = []
+                self.tag_clouds[field_id] = [tag for [tag, _] in field[
+                    'summary']['tag_cloud']]
+                self.term_analysis[field_id] = {}
+                self.term_analysis[field_id].update(
+                    field['term_analysis'])
+            if field['optype'] == 'items':
+                self.items[field_id] = []
+                self.items[field_id] = [item for item, _ in \
+                    field['summary']['items']]
+                self.item_analysis[field_id] = {}
+                self.item_analysis[field_id].update(
+                    field['item_analysis'])
+            if categories and field['optype'] == 'categorical':
+                self.categories[field_id] = [category for \
+                    [category, _] in field['summary']['categories']]
+            if numerics and self.missing_numerics and \
+                    field['optype'] == 'numeric':
+                self.numeric_fields[field_id] = True
+
 
     def uniquify_varnames(self, fields):
         """Tests if the fields names are unique. If they aren't, a
@@ -188,3 +281,65 @@ class ModelFields(object):
             LOGGER.error("Failed to read input data in the expected"
                          " {field:value} format.")
             return ({}, []) if add_unused_fields else {}
+
+    def get_unique_terms(self, input_data):
+        """Parses the input data to find the list of unique terms in the
+           tag cloud
+
+        """
+        unique_terms = {}
+        for field_id in self.term_forms:
+            if field_id in input_data:
+                input_data_field = input_data.get(field_id, '')
+                if isinstance(input_data_field, basestring):
+                    case_sensitive = self.term_analysis[field_id].get(
+                        'case_sensitive', True)
+                    token_mode = self.term_analysis[field_id].get(
+                        'token_mode', 'all')
+                    if token_mode != TM_FULL_TERM:
+                        terms = parse_terms(input_data_field,
+                                            case_sensitive=case_sensitive)
+                    else:
+                        terms = []
+                    full_term = input_data_field if case_sensitive \
+                        else input_data_field.lower()
+                    # We add full_term if needed. Note that when there's
+                    # only one term in the input_data, full_term and term are
+                    # equal. Then full_term will not be added to avoid
+                    # duplicated counters for the term.
+                    if token_mode == TM_FULL_TERM or \
+                            (token_mode == TM_ALL and terms[0] != full_term):
+                        terms.append(full_term)
+                    unique_terms[field_id] = get_unique_terms(
+                        terms, self.term_forms[field_id],
+                        self.tag_clouds.get(field_id, []))
+                else:
+                    unique_terms[field_id] = [(input_data_field, 1)]
+                del input_data[field_id]
+        # the same for items fields
+        for field_id in self.item_analysis:
+            if field_id in input_data:
+                input_data_field = input_data.get(field_id, '')
+                if isinstance(input_data_field, basestring):
+                    # parsing the items in input_data
+                    separator = self.item_analysis[field_id].get(
+                        'separator', ' ')
+                    regexp = self.item_analysis[field_id].get(
+                        'separator_regexp')
+                    if regexp is None:
+                        regexp = ur'%s' % re.escape(separator)
+                    terms = parse_items(input_data_field, regexp)
+                    unique_terms[field_id] = get_unique_terms(
+                        terms, {},
+                        self.items.get(field_id, []))
+                else:
+                    unique_terms[field_id] = [(input_data_field, 1)]
+                del input_data[field_id]
+
+        if hasattr(self, "categories") and self.categories:
+            for field_id in self.categories:
+                if field_id in input_data:
+                    input_data_field = input_data.get(field_id, '')
+                    unique_terms[field_id] = [(input_data_field, 1)]
+                    del input_data[field_id]
+        return unique_terms
